+++ assets/js/csp/functions/MarkDownInputPreview.js
@@ -1,0 +1,18 @@
+/**
+ * MarkDownInputPreview() function JS
+ *
+ * @since 12.5
+ *
+ * @package RosarioSIS
+ */
+
+csp.functions.markDownInputPreview = {
+	preview: function() {
+		MarkDownInputPreview(this.dataset.id);
+	},
+	onEvents: function() {
+		$('.md-preview .tab').on('click', csp.functions.markDownInputPreview.preview);
+	}
+}
+
+$(csp.functions.markDownInputPreview.onEvents);