+++ functions/Inputs.php
@@ -632,6 +632,8 @@ function TinyMCEInput( $value, $name, $title = '', $extra = '' )
  */
 function MarkDownInputPreview( $input_id )
 {
+	static $js_included = false;
+
 	if ( ! $input_id )
 	{
 		return false;
@@ -639,15 +641,21 @@ function MarkDownInputPreview( $input_id )
 
 	ob_start();
 
+	if ( ! $js_included )
+	{
+		// @since 12.5 CSP remove unsafe-inline Javascript
+		?>
+		<script src="assets/js/csp/functions/MarkDownInputPreview.js?v=12.5"></script>
+		<?php
+
+		$js_included = true;
+	}
+
 	?>
 	<div class="md-preview">
-		<a href="#" onclick="<?php echo AttrEscape( 'MarkDownInputPreview(' .
-			json_encode( $input_id ) .
-			'); return false;' ); ?>" class="tab disabled"><?php echo _( 'Write' ); ?></a>
+		<a href="#!" data-id="<?php echo AttrEscape( $input_id ); ?>" class="tab disabled"><?php echo _( 'Write' ); ?></a>
 
-		<a href="#" onclick="<?php echo AttrEscape( 'MarkDownInputPreview(' .
-			json_encode( $input_id ) .
-			'); return false;' ); ?>" class="tab"><?php echo _( 'Preview' ); ?></a>
+		<a href="#!" data-id="<?php echo AttrEscape( $input_id ); ?>" class="tab"><?php echo _( 'Preview' ); ?></a>
 
 		<a href="https://gitlab.com/francoisjacquet/rosariosis/wikis/Markdown-Cheatsheet" title="<?php echo AttrEscape( _( 'Mastering MarkDown' ) ); ?>" target="_blank" class="md-link">
 			<img class="button" src="assets/themes/<?php echo Preferences( 'THEME' ); ?>/btn/md_button.png" alt="<?php echo AttrEscape( _( 'Mastering MarkDown' ) ); ?>">